Multi-threaded program, starting to work
This version can be inserted between various clients and linHPSDR
in the CAT stream. GUI buttons will command the SteppIR but programs
such as WSJT-X changing the frequency won't cause a change to the
SteppIR controller yet: Serial communication to the SteppIR controller
is too slow and messes up the CAT communication. I'll need to put the
serial comms into a separate thread to fix this.