Added a 4th thread for serial comms
Created a 4th thread to do serial communications. Serial was too slow
to be able to do it inside the other threads, so had to implement in a
separate thread. The program now changes the SteppIR SDA-100 controller's
frequency if AUTOTRACK is on and the CLIENT changes the frequency via
the CAT port.