Disabling the RadioQueryLoop thread for now.
The RadioQueryLoop thread, if run in conjunction with a CAT controller
connected to the listening port, will interfere with the CAT controller.  I've
disabled the thread for now. The correct implementation for this thread would
be to have it started if and only if there is nothing connected to the
listening port, and to kill the thread when something connects.